ocrPipeline/requirements.txt
pytesseract-0.3.10
pdf2image-1.16.3
pip install pillow